Fix the tests: cast u8 as usize because of d27a1955dbefa6c6df807b2ab68dde55b00236b5

+++ src/network/mod.rs
@@ -192,7 +192,7 @@ mod tests{
         new_network_test(8, 1);
     }
 
-    fn new_network_test(network_size: usize, initiated_connections: usize) {
+    fn new_network_test(network_size: u8, initiated_connections: usize) {
         let network = Network::new(network_size, initiated_connections);
 
         let global_number_of_received_messages = Arc::new(AtomicUsize::new(0));
@@ -211,8 +211,8 @@ mod tests{
             }
         }, Duration::from_secs(5));
 
-        assert_eq!(network_size * 2 * initiated_connections, connections_established.load(Ordering::Relaxed));
-        assert_eq!(network_size * 2 * initiated_connections, global_number_of_received_messages.load(Ordering::Relaxed));
+        assert_eq!(network_size as usize * 2 * initiated_connections, connections_established.load(Ordering::Relaxed));
+        assert_eq!(network_size as usize * 2 * initiated_connections, global_number_of_received_messages.load(Ordering::Relaxed));
         assert!(notified_of_start.load(Ordering::Relaxed));
     }
 }
